move flash methods to flash header/source
